Remove Stoppable, move store code into lib
