+++ build.gradle
@@ -4,6 +4,7 @@ buildscript {
         kotlin_coroutines_version = '1.2.1'
         dokka_version = '0.9.18'
         jvm_version = '1.8'
+        mockk_version = '1.9.3'
         detekt_version = '1.0.0-RC14'
         // Version to publish
         watchable_version = '0.6.10'
@@ -56,6 +57,7 @@ dependencies {
     testImplementation 'junit:junit:4.12'
     testImplementation 'com.nhaarman.mockitokotlin2:mockito-kotlin:2.1.0'
     testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:$kotlin_coroutines_version"
+    testImplementation "io.mockk:mockk:$mockk_version"
 
     detektPlugins "io.gitlab.arturbosch.detekt:detekt-formatting:$detekt_version"
 }
