+++ sample/src/main/kotlin/store/FileStore.kt
@@ -16,6 +16,8 @@
 
 package store
 
+import io.gladed.watchable.store.Store
+import io.gladed.watchable.store.cannot
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.flow.Flow
