+++ src/main/kotlin/io/gladed/watchable/store/MemoryStore.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package store
+package io.gladed.watchable.store
 
 import io.gladed.watchable.util.guarded
 import kotlinx.coroutines.FlowPreview
