+++ src/test/kotlin/store/HoldingStoreTest.kt
@@ -1,3 +1,21 @@
+/*
+ * (c) Copyright 2019 Glade Diviney.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package store
+
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.mockk
@@ -12,19 +30,21 @@ import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 import kotlinx.coroutines.test.TestCoroutineScope
-import model.Bird
 import org.junit.Assert.assertEquals
 import org.junit.Assert.fail
 import org.junit.Before
 import org.junit.Test
-import store.Cannot
-import store.Hold
-import store.HoldingStore
-import store.Store
-import test.runTest
+import io.gladed.watchable.store.Cannot
+import io.gladed.watchable.store.Hold
+import io.gladed.watchable.store.HoldingStore
+import io.gladed.watchable.store.Store
+import runTest
+import java.util.UUID
 
 @UseExperimental(ObsoleteCoroutinesApi::class, ExperimentalCoroutinesApi::class)
 class HoldingStoreTest {
+    data class Bird(val id: String = UUID.randomUUID().toString(), val name: String)
+
     private val robin = Bird(name = "robin")
 
     private val rootStore = mockk<Store<Bird>>(relaxUnitFun = true)
